y1

+++ app/Http/Controllers/Admin/CompanyController.php
@@ -146,7 +146,7 @@ class CompanyController extends Controller
             'up_companyName' => 'required',
             'up_telephoneNo' => 'required',
             'up_mailingAddress' => 'required|unique:company,company.mailingAddress'.$request->up_mailingAddress,
-            'file' => 'image|mimes:jpg,jpeg,png,gif,svg|max:2048',
+            'filenew' => 'image|mimes:jpg,jpeg,png,gif,svg|max:2048',
         ]);
 
         // $companyIDUp=(int)$request->companyID;
@@ -166,21 +166,34 @@ class CompanyController extends Controller
                      }
                 }
        }
-        if ($request->hasFile('filenew') && $request->file('filenew') != '') {
-            if(!empty($companyArrayUp[$v]['file'])){
-            $imagePath = public_path('CompanyLogo/'.$companyArrayUp[$v]['file'][0]['filename']);
-            if(File::exists($imagePath)){
-                unlink($imagePath);
+
+        $photo_name='';
+        $original_name='';
+        $size='';
+        $photo_path='';
+
+        if($request->filenew != null){
+            if ($request->hasFile('filenew') && $request->file('filenew') != '') {
+                if(!empty($companyArrayUp[$v]['file'][0]['filename'])){
+                    $imagePath = public_path('CompanyLogo/'.$companyArrayUp[$v]['file'][0]['filename']);
+                    if(File::exists($imagePath)){
+                        unlink($imagePath);
+                    }
+                }
+                $files = $request->file('filenew');
+                $ImageUpload = Image::make($files);
+                $originalPath = 'CompanyLogo/';
+                $ImageUpload->save($originalPath.time().$files->getClientOriginalName());
+                $photo_path = 'CompanyLogo/'.time().$files->getClientOriginalName();
+                $photo_name = time().$files->getClientOriginalName();
+                $original_name = $files->getClientOriginalName();
+                $size = $request->file("filenew")->getSize();
             }
-            }
-            $files = $request->file('filenew');
-            $ImageUpload = Image::make($files);
-            $originalPath = 'CompanyLogo/';
-            $ImageUpload->save($originalPath.time().$files->getClientOriginalName());
-            $photo_path = 'CompanyLogo/'.time().$files->getClientOriginalName();
-            $photo_name = time().$files->getClientOriginalName();
-            $original_name = $files->getClientOriginalName();
-            $size = $request->file("filenew")->getSize();
+        }else{
+            $photo_name=$companyArrayUp[$v]['file'][0]['filename'];
+            $original_name=$companyArrayUp[$v]['file'][0]['Originalname'];
+            $size=$companyArrayUp[$v]['file'][0]['filesize'];
+            $photo_path=$companyArrayUp[$v]['file'][0]['filepath'];
         }
 
        $companyArrayUp[$v]['companyName']=$request->up_companyName;
@@ -244,8 +257,26 @@ class CompanyController extends Controller
     }
 
     public function updateUserCompany(Request $request){
+        $email=$request->email;
+        $result = Company::where('companyID',1 )->first();
+        $companyArray=$result->company;
+        $arrayLength=count($companyArray);
+        $i=0;
+        $v=0;
+        for ($i=0; $i<$arrayLength; $i++){
+                $id=$result->company[$i];
+            
+                    foreach ($id as $value){
+                        if($value==$email){
+                            $v=$i;
+                        }
+                    }
+        }
+        $compidupdate = $companyArray[$v]['_id'];
+
         $user = Auth::user();
-        $user->companyID = $request->companyId;
+        $user->companyName = $compidupdate;
+        $user->companyID = 1;
         $user->save();
         $data = [
             'success' => true,
